GEODE-8772: Assign TCP Server test ports in test JVM
Make TcpServerProductVersionDUnitTest assign the locator port in the
test JVM rather than in child VMs.

An upcoming change to AvailablePortHelper will improve how it assigns
ports to better support tests running in parallel outside of docker.  A
child VM running an older version of Geode (as happens in this test)
will have an out-of-date implementation of AvailablePortHelper, and so
may assign a port that is in use by another test running in parallel.