GEODE-9228: Fix possibly batches with incomplete transaction events sent when stopping gateway sender.
When stopping a gateway sender having group-transaction-events
set, there is a grace period in which events
received that belong to transactions are not dropped
if there are events in the queue belonging
to that transaction. This is done to make
sure that the queue does not contain incomplete
transactions.

Given that when the parallel gateway sender
creates a batch with events from the queue,
it removes them from the queue as it adds them
to the batch, it is possible that a transaction event
received by a gateway sender during
the stopping grace period
does not find events belonging to that transaction
in the queue because they have been added to
a current batch and therefore that event will
be dropped, causing that an incomplete transaction
will be finally stored in the queue/batch.
In order to solve the issue, when receiving a
transaction event during the stopping grace period,
the gateway sender must look for events belonging
to that transaction, not only in the queue
but also in the peekedEvents from the queue
(those events retrieved from the queue and
put in batches).

Also, in order to avoid flakyness of the distributed
test cases that test that batches with incomplete
transactions are not sent when stopping the
gateway sender, the number of retries
to get the events from the queue/batches
has been set to high number in the tests by means of
a new setter in GatewaySender which has
also been published as a gfsh and
cache.xml parameter (get-transaction-events-from-queue-retries).