GEODE-8971: Fix possible incomplete transactions sent when stopping gateway sender
A bug was found in the SerialGatewaySenderQueue.hasEventsMatching() function
that provoked that sometimes incomplete transactions were sent
in batches when group-transaction-events is true, right
before the gateway sender is stopped.

The bug consisted in an error in the mentioned
function that when checking if there were
events in the queue belonging to a transaction did
not check the event with the currentKey. Instead
it started with the currentKey+1 until the
tailkey was reached.
This provoked that some transaction events received in the
gateway sender during the grace period before it is stopped
were not queued even though there were events in the queue
for the same transaction.

A couple of small refactorings have also been made
in the SerialGatewaySenderQueue class.