GEODE-9229: Avoid possible duplicate events sent with group-transaction-events
When group-transaction-events is set to true
the SerialGatewaySenderQueue class stores those events
peeked from the queue to complete transactions
for a batch inside the extraPeekedIds member. This is done
so that, when new events are peeked from the queue
by peekAhead() to be put in a batch, the
events in extraPeekedIds are skipped. Otherwise
those events will be sent several times.
When the remove method of SerialGatewaySenderQueue
is called, the peekedIds member and the head key
of the queue are updated and events from extraPeekedIds
are removed if necessary.
Given that the peekAhead() method first gets the
currentKey (that depends on the peekedIds and the headkey)
and then checks if the event to be removed is in extraPeekedIds,
it could be that in between those two calls,
remove() is called from another thread, causing that
peekAhead() returns an event that
belonged to extraPeekedIds when the currentKey was
obtained but was deleted from extraPeekedIds by remove()
before it was checked that the event belonged to
extraPeekedIds.
This provokes that an event already peeked is returned
again by peekAhead() leading to duplicate events
sent by the gateway sender.

The solution proposed to solve this problem is to make
synchronized the peekAhead() function when group-transaction-events
is set given that remove() is already synchronized. This will
prevent the race condition described above.