GEODE-9256: Adding a jmh test of the Object2ObjectOpenCustomHashmapWithCursor
The test compares the performance of the cursor with the performance
of an iterator. Although the cursor is about an order of magnitude
slower than an iterator, the iteration time still looks like it's
O(1) - eg the time to iterate per entry does not go up with the
number of entries.

Tweaking the name of the elementAtHashesTo method, based on feedback from the
last PR. Trying to make that code a little more readable.