New solution:
Normal put operation (transaction is not used) is distributed to adjunct
member (only host sub. queue, and not bucket) using seperate
partition.PutMessage which indicates that it is notificationOnly, and
also sets TailKey to -1. Adjunct member will not store event in queue,
because TailKey is always expected to be set in case of parallel
gateway-sender. Members that host secondary bucket will receive the
UpdateOperation message with valid TailKey value.

So this new solution implements the similar approach to what is done in
case of normal put operation. TxCommitMessage with TailKey set to -1
will be sent only to adjunct member that host subscription queue, and not
secondary bucket. Other members will get valid TailKey value. Also,
member that host both secondary bucket and subscription queue will get
valid TailKey value.