Several fixes, new test cases and refactorings.
Handle ConnectionDestroyed and Server disconnection exceptions
that could appear when there is a socket-read-timeout configured in the gateway
sender.

Fix cancel command that was not stopping commands the second time it
was launched because there were several threads with the same name in the pool.

Fix calculation of sleep time to adjust to maxRate.

Added unit tests for replicateRegion function.