GEODE-8772: Disable HTTP service in Client Server TX test
PROBLEM

`ClientServerTransactionFailoverWithMixedVersionServersDistributedTest`
launches a locator with HTTP service enabled on the default port (7070).
If these tests run in parallel outside of docker, each locator tries to
bind the HTTP service to port 7070, and all but the first will fail.

This happens even though the tests do not need the HTTP service.

THIS COMMIT

Starts the locators with HTTP service disabled.

Authored-by: Dale Emery <demery@vmware.com>