GEODE-8772: Fix ClusterComms test port conflicts
Change `ClusterCommunicationsDUnitTest.performARollingUpgrade()` to
disable the HTTP service.

BACKGROUND

I am working on a project to allow Geode tests to run in parallel
outside of Docker. Running in parallel outside of Docker requires that
tests not start services using default ports.

This commit prepares for those changes.

PROBLEMS

When `ClusterCommunicationsDUnitTest.performARollingUpgrade()` upgrades
the locator, it starts the new locator with HTTP service enabled on the
default port (7070). If another test running in parallel also uses this
default port, at least one will throw an exception when it attempts to
bind.

THIS COMMIT

Change `ClusterCommunicationsDUnitTest` to start the upgraded locator
with HTTP service disabled. The test does not use the HTTP service.