GEODE-9340: revert upgrade of jetty version
The new versions of jetty (after 9.4.35.v20201120) have some yet
undiagnosed bug that causes a significant performance regression,
especially when doing operations on small amounts of data, like longs.
The data from this investigation, as well as notes and CI failure links,
can all be found in the jira ticket. Rolling jetty back to a previous
version is a short-term fix, but a long term solution is going to have
to come from Eclipse.