Revert "GEODE-8971: Add grace period when stopping gateway sender with group-… (#6052)" (#6634)
This reverts commit 841fa06c7b34916c09df920b7029974a78255cd0.

The reason is that this commit creates problems with ongoing
put operations during the grace period (very long response
times) and also, if operations rate is high, it could prevent
that the gateway sender is stopped.

Put operations can be very slow during the grace period
because they need to traverse the gateway sender queue
to find events with the same transactionId. If the queue
size is big and events are evicted to disk, the time to
process a request can be unacceptably long.
At the same time, this can provoke that the gateway sender
is never stopped because it is blocked trying to get
a write lock and the lock is held by ongoing operations
for a long time.