GEODE-9340: move benchmark baseline
The new baseline is the commit that caused a slight performance
regression in our most sensitive tests. There is a negligible
performance difference between the previous baseline of 1.13.3 and the
commit before the commit that caused the regression (commit with SHA
217be41f087072026ea3b411b89107700edfb5f7). This means that moving the
baseline will not ignore a bunch of other performance changes.