GEODE-9421: Remove BatchRemoval NPE logs from JUnit tests
ParallelAsyncEventQueueImplTest and ParallelGatewaySenderImplTest classes were mocking InternalRegion
but were not providing an implementation for cache.getCancelCriterion().
This led to a null pointer exception in the BatchRemoval thread code that,
even though it did not affect the tests, provoked that a log message with
a null pointer exception was printed.

The solution consists of adding a response in the mock using when() for the cache.getCancelCriterion() method.