Optimize sizing of RedisHash, RedisSet, and RedisSortedSet.
Removed SizeableObjectSizer and made the Sizeable redis classes abstract.
They now require subclasses that implement sizeKey, sizeValue, or sizeElement.

Instead of having hard coded constants stating what the size of a redis base class is,
the code now uses JvmSizeUtils.sizeClass to statically compute the class size overhead.
This allows the size to be correct for different jvm configs.
Also simplified the sizing logic to use int and if needed use casts instead of narrow.
Narrow was more expensive and would still cause problems as the size goes back down toward zero.
The sizing of OrderStatisticsTree was simplified some more since it did not need an extra int field
since we do not keep track of its element sizes.
Also the size of the hashing strategy object is no longer accounted for since it is a singleton.