GEODE-9483: Make StartLocatorCommandIntegrationTest not start unneeded locator
Two of the tests in `StartLocatorCommandIntegrationTest` launch a locator
process that binds to the default HTTP service port (7070). This can
interfere with the tests that explicitly verify that the system uses the
correct HTTP service port by default.

Neither test actually needs the launched locator for any purpose.
Neither test makes a single assertion about the locator. When the tests
exit, they leave the locator running. In fact, the locator typically
finishes launching well after the tests have all exited.

Rather than fix the port assignment, this commit changes the tests so
that they do not actually launch a locator. Two of the tests in this
class already do this.