start up Docker containers using available ports
used the AvailablePortHelper to generate a list of available ports for
the docker containers to use.
for each docker container there is the string "#PORT_n" in
redis-cluster-compose.yml that can be replaced with the available port
in the NativeRedisClusterTestRule. the edited yaml is passed into the
constructor for the Docker container, so each test container will have
its own available port without collisions.