GEODE-9408: Avoid duplicate events sent by Serial Gateway Sender when group-transaction-events is true (#6663)
GEODE-9408: Avoid duplicate events sent by Serial Gateway Sender

The logic for removing events from the Serial Gateway Sender queue
when an ack for a batch is received from the gateway receiver
is implemented by the remove() method of the SerialGatewaySenderQueue class.
When the group-transaction-events setting is activated, this logic,
needs to act differently if the event was an event peeked to complete
a transaction in a batch than if it is an event peeked with the peekAhead method.
In the first case, the head key and lastDispatched member must not be
updated with the key of the event because that could provoke that
events from the queue not yet retrieved are never retrieved.

Nevertheless, the head key and lastDispatched member must be
updated for a key peeked to complete a transaction (extraPeekedId)
when the key with the previous id is removed.
This was not being done in some cases by the remove() method and it
has been corrected in this PR.

Several DUnit test cases have been added to test this fix
in a Geode WAN system with a Serial GatewaySender with
group-transaction-events set to true. In the test cases, it is verified that
no duplicated events are sent by means of checking that incomplete transactions
are not sent.

Increase default number of retries