GEODE-9497: create new hash map class with fewer fields (#6768)
Bytes2ObjectOpenHashMap is now used by RedisHash and RedisSortedSet for its map.
This saves around 20 bytes per "hash" or "zset" instance.
Bytes2ObjectOpenHashMap is derived from Object2ObjectOpenCustomHashMap so if you see
some code in it that you question, compare it to the original. The class comment on Bytes2ObjectOpenHashMap describes the changes made to it. The goal was to keep the code as
close to the original, but to make some changes to save on memory. Some cosmetic changes were
also made using intellij automated cleanup (for example adding curly braces, getting rid of extra parens).

Some new top level methods were added to allow users to avoid the overhead of creating a new instance
when calling keySet(), entrySet(), or values(). These methods use to cache their result but no longer do to save on memory footprint. These new methods are named "fast*" and use the same pattern of some fast iteration code from the original Object2ObjectOpenCustomHashMap.