GEODE-9347: optimize pubSub (#6814)
* reworked PubSub data structures.
Each server keeps track of all its local subscriptions in a SubscriptionManager.
The SubscriptionManager keeps track of Clients subscribed in a ClientSubscriptionManager.

Each Client keeps track of its subscription names in a set.
Client now keeps track of channel and patterns it has subscribed to making it easier to check. This will improve performance of non-pubsub ops when a server has existing subscriptions.

The internal Subscriptions class now uses maps instead of lists. Concurrent maps are used and the synchronization in the pubsub internal implementation has been removed.
* Some of the messages have been optimized by sending a canonical byte[] instead of converting a String to a byte[] for every message.
* If many clients subscribe to the same pattern, publish will now only need to check that pattern once. The older code checked once per client.
* The Subscription class is now smaller since some of its fields are now keys in the Manager maps.
*Now we have a single channel shutdown listener per Client.
Before we had one for every Subscription and unsubscribe did
not clean them up. If the same client subscribed and unsubscribed this would result in a memory leak.