GEODE-9585: adding support for the KEYSLOT command
Adding support for the keyslot command. refactoring the RedisKey code a little
bit to move the slot calculation into a static method that can be used and
tested separately. Changing RedisKey.crc16 to slot, because that is only thing
we are using the crc16 field for.

Changing the implementation of slotForKey to more closely mirror the logic of
the native redis keyHashSlot method. This reproduces the same behavior as
native redis with malformed hashtags with unmatched parens.

Adding more test cases including a junit quick check case.

Co-authored-by: Eric Zoerner <zoernere@vmware.com>