GEODE-9581: Fix possible sending of duplicate events
When group-transaction-events is set for a serial
gateway sender there is a small chance that duplicated
events are sent by the sender.

The reason behind is a bug in the serial gateway sender.

When the remove() method of the SerialGatewaySenderQueue
was invoked for an event that was peeked to complete
a transaction, it was also removed from the extraPeekedIds
structure.
This provoked that events already peeked, were peeked
again by SerialGatewaySenderQueue.peekAhead() when
they should not have been.

A change has been made in the remove() method so
that events that were peeked to complete a transaction
are only removed from the extraPeekedIds when a call
to remove() has been made previously for that event
(the events is in extraPeekedIdsRemovedButPreviousIdNotRemoved)
and the previous event is the lastDispatchedKey.