Revert "Revert "GEODE-9607: prevent pubsub from running server out of memory (#6880)" (#6921)"
This reverts commit 45383f0e562f9ba738858fe91fd38fba9e9a1023.

1. publish no longer creates extra netty listener objects
2. publish now writes directly to the netty ByteBuf instead
of delaying it using a RedisResponse object.
3. a publisher thread will now loop around and publish another
batch instead of handing it off to another thread

the batch is now optimized by conflating consecutive publish requests
on the same channel. This allows them to be written to netty as
a single message instead of multiple messages

Added a new PubSubConcurrentDUnitTest
If a publish is going to be sent to multiple netty channels,
it now writes the message once to a ByteBuf and shares that
ByteBuf with each channel.
The batch of publish requests is now optimized by combining
consecutive messages to the same channel.
Pattern publish now caches which subscribers match a particular
publish channel name so that the glob matching does not need
to be done on each publish.
Some of the foreach support has been dropped from the pubsub
manages in favor of getting a Collection of subscriptions so
that the publish channel writes can be grouped efficiently.