GEODE-9650: LOLWUT command (#6908)
This implements a version of the Redis LOLWUT command, which is supposed to (a) algorithmically generate computer art, and (b) output the program version. Optionally, the LOLWUT command can take arguments to affect the output.

This implementation generates arbitrary-size mazes. By default, it generates a 40-cell-wide maze of height 10, but given arguments it can generate a valid maze of any width and height (up to 1Kx1M), while consuming a small and fixed amount of memory. Below the maze, it outputs the Geode version.

Co-authored-by: Ray Ingles <ringles@vmware.com>