GEODE-9536: Make PingOp test more patient
PROBLEM

On some versions of Windows (such as the version we use in CI), the
granularity of `System.currentTimeMillis()` is ~15ms.

`PingOpDistributedTest.regularPingFlow()` executes two pings and
asserts that the resulting heartbeat timestamp to has increased from one
ping to the next.

The test's two pings often occur so rapidly that the heartbeat
timestamps appear to be the same, causing the assertion to fail.

SOLUTION

Change the test to use `await()` to repeat the pings until the heartbeat
timestamp increases.

NOTE

This prepares for an upcoming PR to run all distributed tests on
Windows.