Fixing the archiving of benchmark results (#7000)
We were using the wrong directory name when trying to create
a benchmark archive. It was looking for a single directory in the format
benchmarks-apache-develop-main-benchmark-base-277-89433-base, but there
are actually multiple directories in the format
benchmarks-apache-develop-main-benchmark-base-277-89433-base-output-10-13-2021-18-26-22-apache-develop-main-benchmark-base-277-89433-base
as well as a text file.

Archiving the whole results directory. Also, including a tabular
dump of the results in the archive for potentially easier parsing.