GEODE-9562: add system property for redis region name (#6960)
Adds a system property (geode-for-redis-region-name) that can be set to
any non-null, non-empty string. By default the data region for geode for
redis is GEODE_FOR_REDIS. If the system property value is not valid, the
default will quietly be used. The region name will be used in the RegionProvider
and in authentication. Supports both geode. and gemfire. prefixes for the 
system property name

other changes:
- remove comments that say Geode for Redis is experimental (it isn't)