Fixing a race condition in PubSubIntegrationTest
Fixing a race condition in ensureOrderingOfPublishedMessages. This test was
expecting the client to receive messages from the server after calling publish,
but the events are processed asynchronously on the client. Changing the test to
wait for the messages to arrive.

Backporting the develop version of this test resulted in new failures, so just
fixing the one race condition we're hitting in support/1.14 for this test.