ADPPRG-62536: Temporary fix to gateway sender stop getting stuck
When stopping a parallel gateway sender it has been
sometimes observed that the command hangs.
The command gets stuck in a call to get the size of the queue
to decide if it must log a warning message telling
that the gateway sender is being stopped with
actively queued data.

The root cause for this call getting stuck
is still under investigation.

In the mean time, the call to get the size of the queue
and the log message have been removed to
avoid the hanging of the command.

geode-core/src/main/java/org/apache/geode/internal/cache/wan/AbstractGatewaySenderEventProcessor.java