GEODE-9781: clean up radish sys props (#7080)
* all geode-for-redis system property names are now defined in RedisProperties.
A check is now done that the property name has the required prefix.
Warnings are logged if the system property is set to a value that will not be used.

* AuthIntegrationTest was trying to test with a non-default region name by setting
the system property "REDIS_REGION_NAME_PROPERTY" in setupCacheWithRegionName.
But this actually did nothing and geode still used the default. The problem was that
the constant was not actually the name of the system property. It now is and this test
should now be using a non-default region name.

* if the bucket sys prop is greater than 16384 it will now log a warning and use the default number of buckets. It used to fail the redis server startup.

* RedisPropertiesTest now uses RestoreSystemProperties rule